Refactor ETag cache mode into separate plugin
